requirements.txt
tensorflow-cpu==2.4.1
Flask==1.1.1
Flask-Cors==3.0.8
Pillow==7.0.0
opencv-python==4.5.3.56
